Move bucket setup helpers to r2.py; fix local public policy and website paths
Consolidates enable_r2_dev_public, set_public_policy (now via mc CLI), apply_cors,
and ensure_bucket into r2.py. Uses mc anonymous set for MinIO local dev instead of
the unsupported PutBucketPolicy. Updates website metric names to match the new
table-based export paths.
