+++ scripts/setup_local.py
@@ -19,7 +19,13 @@ ROOT = Path(__file__).parent.parent
 sys.path.insert(0, str(ROOT))
 
 from pipeline.config import PipelineConfig
-import pipeline.r2 as R2
+from pipeline.r2 import (
+    create_inboxes,
+    make_client,
+    ensure_bucket,
+    make_web_client,
+    set_public_policy
+)
 
 
 def ensure_minio(endpoint: str) -> None:
@@ -39,21 +45,17 @@ def ensure_minio(endpoint: str) -> None:
     sys.exit(1)
 
 
-def ensure_bucket(r2, bucket: str) -> None:
-    try:
-        r2.client.head_bucket(Bucket=bucket)
-    except ClientError:
-        r2.client.create_bucket(Bucket=bucket)
-        print(f"✓ Created bucket '{bucket}'")
-    print(f"✓ Bucket '{bucket}' ready")
-
 
 def main() -> None:
     config = PipelineConfig.load(ROOT / "config" / "test.toml", ".env.local.example")
+    r2 = make_client(config)
+    web_r2 = make_web_client(config)
 
-    r2 = R2.make_client(config)
     ensure_minio(config.endpoint_url)
     ensure_bucket(r2, config.r2_bucket_name)
+    ensure_bucket(web_r2, config.web_bucket_name)
+    set_public_policy(config.endpoint_url,config.web_bucket_name)
+    create_inboxes(r2)
 
 
 if __name__ == "__main__":
