+++ scripts/setup_r2.py
@@ -19,16 +19,20 @@ from __future__ import annotations
 import json
 import sys
 from pathlib import Path
-from urllib.parse import urlparse
 
 ROOT = Path(__file__).parent.parent
 sys.path.insert(0, str(ROOT))
 
-import httpx
-
 from pipeline.config import PipelineConfig
-from pipeline.r2 import make_client, make_web_client, upload_bytes, exists
-from pipeline.main import ALL_JOBS
+from pipeline.r2 import (
+    apply_cors,
+    enable_r2_dev_public,
+    make_client,
+    make_web_client,
+    set_public_policy,
+    create_inboxes,
+    ensure_bucket,
+)
 
 
 def main() -> None:
@@ -41,80 +45,12 @@ def main() -> None:
 
     ensure_bucket(r2, config.r2_bucket_name)
     ensure_bucket(web_r2, config.web_bucket_name)
-    if config.runtime_env != "local":
-        enable_r2_dev_public(config.endpoint_url, config.web_bucket_name, config.secrets.cloudflare_api_token)
+    enable_r2_dev_public(config.endpoint_url, config.web_bucket_name, config.secrets.cloudflare_api_token)
     apply_cors(web_r2.client, config.web_bucket_name, cors_rules)
     create_inboxes(r2)
-
     print("✓ R2 setup complete")
 
 
-def ensure_bucket(r2, bucket: str) -> None:
-    from botocore.exceptions import ClientError
-
-    try:
-        r2.client.head_bucket(Bucket=bucket)
-        print(f"· Bucket '{bucket}' found")
-    except ClientError as e:
-        code = e.response["Error"]["Code"]
-        if code in ("404", "NoSuchBucket"):
-            r2.client.create_bucket(Bucket=bucket)
-            print(f"✓ Bucket '{bucket}' created")
-        else:
-            raise
-
-
-def enable_r2_dev_public(endpoint_url: str, bucket: str, api_token: str) -> None:
-    host = urlparse(endpoint_url).hostname or ""
-    account_id = host.split(".")[0]
-
-    url = f"https://api.cloudflare.com/client/v4/accounts/{account_id}/r2/buckets/{bucket}/domains/managed"
-
-    resp = httpx.put(
-        url,
-        headers={
-            "Authorization": f"Bearer {api_token}",
-            "Content-Type": "application/json",
-        },
-        json={"enabled": True},
-        timeout=30.0,
-    )
-    data = resp.json()
-    if resp.status_code == 200 and data.get("success"):
-        enabled = data["result"]["enabled"]
-        domain = data["result"]["domain"]
-        if enabled:
-            print(f"✓ Public r2.dev enabled: https://{domain}")
-        else:
-            print("⚠️ Public access wasn't enabled for some reason")
-    else:
-        print(f"✗ Failed to enable public access: {resp.status_code} {data}")
-
-
-def create_inboxes(r2) -> None:
-    """Create placeholder objects for each source inbox folder."""
-    sources = [
-        "fitbit",
-        "github",
-        "gymgroup",
-        "kindle",
-        "macos_commands",
-        "macos_screentime",
-        "strong"
-    ]
-    for src in sources:
-        placeholder = src + "/.keep"
-        if not exists(r2, placeholder):
-            upload_bytes(r2, placeholder, b"", "application/octet-stream")
-            print(f"  created {src}/")
-
-
-def apply_cors(client, bucket: str, cors_rules: list) -> None:
-    try:
-        client.put_bucket_cors(Bucket=bucket, CORSConfiguration={"CORSRules": cors_rules})
-        print(f"✓ CORS rules applied to '{bucket}'")
-    except Exception as e:
-        print(f"✗ Failed to apply CORS rules: {e}")
 
 
 if __name__ == "__main__":
